添加退出，邀请多人功能，去掉国家代码前缀

+++ Main/Tools/Tool.m
@@ -532,7 +532,7 @@
         
     }else{
         
-        return [NSString stringWithFormat:@"%ld/%ld/%ld",(long)datemonth,dateday,dateyear];
+        return [NSString stringWithFormat:@"%ld/%ld/%ld",(long)datemonth,(long)dateday,(long)dateyear];
         
     }
     //
